+++ backend/controller/auth.go
@@ -1,0 +1,64 @@
+package controller
+
+import (
+	"laundry-ledger/db"
+	"laundry-ledger/helper"
+	"laundry-ledger/model"
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+)
+
+func Signup(c *gin.Context) {
+	var user model.User
+	if err := c.ShouldBindJSON(&user); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	// Save user to database
+	if err := db.DB.Create(&user).Error; err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{"message": "User created successfully"})
+}
+
+func Login(c *gin.Context) {
+	var request struct {
+		Email    string `json:"email"`
+		Password string `json:"password"`
+	}
+	var user model.User
+
+	if err := c.ShouldBindJSON(&request); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	// Check if user exists
+	if err := db.DB.Where("email = ?", request.Email).First(&user).Error; err != nil {
+		c.JSON(http.StatusNotFound, gin.H{"error": "Email not registered"})
+		return
+	}
+
+	// Verify password
+	if request.Password != user.Password {
+		c.JSON(http.StatusUnauthorized, gin.H{"error": "Wrong email or passsword"})
+		return
+	}
+
+	// Generate JWT token upon successful login
+	token, err := helper.GenerateJWT(user.ID)
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to generate token"})
+		return
+	}
+
+	// Return the token in the response
+	c.JSON(http.StatusOK, gin.H{
+		"message": "Login successful",
+		"token":   token,
+	})
+}
